+++ docs/conf.py
@@ -15,7 +15,7 @@
 # relative to the documentation root, use os.path.abspath to make it
 # absolute, like shown here.
 #
-
+import datetime
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
@@ -44,8 +44,8 @@ master_doc = "index"
 
 # General information about the project.
 project = "gRPCAlchemy"
-copyright = "2018, GuangTian Li"
-author = "GuangTian Li"
+copyright = f"{datetime.datetime.now().year}, GuangTianLi"
+author = "GuangTianLi"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout