+++ grpcalchemy/__init__.py
@@ -9,8 +9,8 @@ __author__ = """GuangTian Li"""
 __email__ = "guangtian_li@qq.com"
 __version__ = "0.7.2"
 
-__all__ = ["Blueprint", "Context", "grpcmethod", "DefaultConfig", "Server"]
+__all__ = ["Blueprint", "Context", "grpcmethod", "DefaultConfig", "Server", "Streaming"]
 
-from .blueprint import Blueprint, Context, grpcmethod
+from .blueprint import Blueprint, Context, grpcmethod, Streaming
 from .config import DefaultConfig
 from .server import Server